+++ packages/flagship/src/lib/modules/android/react-native-camera.ts
@@ -1,11 +1,11 @@
-import * as path from '../path';
-import * as fs from '../fs';
-import { Config } from '../../types';
+import * as path from '../../path';
+import * as fs from '../../fs';
+import { Config } from '../../../types';
 import {
   logInfo
-} from '../../helpers';
+} from '../../../helpers';
 
-export function android(configuration: Config): void {
+export function postLink(configuration: Config): void {
   // Add dependencies to /android/app/build.gradle and replace compile command with implementation
   // command due to Gradle 3 changes
   let gradleAppBuild = fs.readFileSync(path.android.gradlePath(), { encoding: 'utf8' });
