+++ packages/flagship/src/lib/modules/android/react-native-leanplum.ts
@@ -1,8 +1,7 @@
-import * as path from '../path';
-import * as fs from '../fs';
-import * as pods from '../cocoapods';
-import { Config } from '../../types';
-import { logInfo } from '../../helpers';
+import * as path from '../../path';
+import * as fs from '../../fs';
+import { Config } from '../../../types';
+import { logInfo } from '../../../helpers';
 
 const kRepository = `maven { url 'https://repo.leanplum.com/' }`;
 
@@ -11,7 +10,7 @@ const kRepository = `maven { url 'https://repo.leanplum.com/' }`;
  *
  * @param {object} configuration The project configuration.
  */
-export function android(configuration: Config): void {
+export function postLink(configuration: Config): void {
   logInfo('patching Android for react-native-leanplum');
 
   // Add the repository to the project repositories.
@@ -22,13 +21,4 @@ export function android(configuration: Config): void {
     'new RNLeanplumPackage()',
     'new RNLeanplumPackage(application)'
   );
-}
-
-/**
- * Patches iOS for the module.
- *
- * @param {object} configuration The project configuration.
- */
-export function ios(configuration: Config): void {
-  pods.add(path.ios.podfilePath(), [`pod "Leanplum-iOS-SDK", '2.1.0'`]);
 }
