+++ packages/flagship/src/lib/modules/android/react-native-navigation.ts
@@ -1,14 +1,14 @@
-import * as path from '../path';
-import * as fs from '../fs';
-import { Config } from '../../types';
-import { logInfo } from '../../helpers';
+import * as path from '../../path';
+import * as fs from '../../fs';
+import { Config } from '../../../types';
+import { logInfo } from '../../../helpers';
 
 /**
  * Patches Android for the module.
  *
  * @param {object} configuration The project configuration.
  */
-export function android(configuration: Config): void {
+export function postLink(configuration: Config): void {
   logInfo('patching Android for react-native-leanplum');
 
   fs.update(
