+++ packages/flagship/src/lib/modules/ios/react-native-adobe-analytics.ts
@@ -1,31 +1,10 @@
-import * as path from '../path';
-import * as fs from '../fs';
-import * as pods from '../cocoapods';
-import { Config } from '../../types';
-import {
-  logError,
-  logInfo
-} from '../../helpers';
+import { Config } from '../../../types';
+import { logError, logInfo } from '../../../helpers';
+import * as path from '../../path';
+import * as fs from '../../fs';
+import * as pods from '../../cocoapods';
 
-export function android(config: Config): void {
-  if (
-    !config.adobeAnalytics ||
-    !config.adobeAnalytics.android ||
-    !config.adobeAnalytics.android.configPath
-  ) {
-    logError('adobeAnalytics.android.configPath must be specified in project config');
-    return process.exit(1);
-  }
-
-  logInfo('Coping ADBMobileConfig.json to Android assets folder');
-  const destination = path.resolve(path.android.assetsPath(), 'ADBMobileConfig.json');
-  const source = path.project.resolve(config.adobeAnalytics.android.configPath);
-  fs.copySync(source, destination);
-
-  logInfo('finished updating Android for react-native-adobe-analytics');
-}
-
-export async function ios(config: Config): Promise<void> {
+export function preLink(config: Config): void {
   if (
     !config.adobeAnalytics ||
     !config.adobeAnalytics.ios ||
